# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentacion",
        "description": " API documentacion tudespachonet.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/productos": {
            "get": {
                "summary": "Ver todos productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "A list of productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Crear nuevo producto",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/productos/{id}": {
            "put": {
                "summary": "Modificar a producto",
                "operationId": "20b20d8e59992eb88abad7dcef3db6d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto modificado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Producto not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a producto",
                "operationId": "4c0c7680428675134afd4d793f934f67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Producto not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/producto/{id}": {
            "get": {
                "summary": "Ver producto por ID",
                "description": "Obtiene los detalles de un producto específico mediante su ID.",
                "operationId": "getProductoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto que se desea obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "producto": {
                                            "$ref": "#/components/schemas/Producto"
                                        },
                                        "nombre_categoria": {
                                            "type": "string",
                                            "example": "Electrónicos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProductoRequest": {
                "title": "ProductoRequest",
                "required": [
                    "nombre",
                    "id_categoria",
                    "descripcion",
                    "precio"
                ],
                "properties": {
                    "nombre": {
                        "type": "string"
                    },
                    "id_categoria": {
                        "type": "integer"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "precio": {
                        "type": "number",
                        "format": "float"
                    },
                    "imagen": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Producto": {
                "title": "Producto",
                "required": [
                    "nombre",
                    "id_categoria",
                    "descripcion",
                    "precio"
                ],
                "properties": {
                    "id": {
                        "description": "ID a producto",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre de producto",
                        "type": "string"
                    },
                    "id_categoria": {
                        "description": "ID categoria",
                        "type": "integer"
                    },
                    "descripcion": {
                        "description": "Descripcion  producto",
                        "type": "string"
                    },
                    "precio": {
                        "description": "precio producto",
                        "type": "number",
                        "format": "float"
                    },
                    "imagen": {
                        "description": "Image URL of the producto",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}